+++ AutomatedLabUnattended/internal/functions/Win/Set-UnattendedWindowsUserLocale.ps1
@@ -24,27 +24,8 @@
 	}
 	else
 	{
-		try
-		{
-			$inputLocale = @((New-WinUserLanguageList -Language $UserLocale).InputMethodTips)
-			$inputLocale += (New-WinUserLanguageList -Language 'en-us').InputMethodTips
-		}
-		catch
-		{
-			Remove-Module -Name International -ErrorAction SilentlyContinue -Force
-			Get-ChildItem -Directory -Path ([IO.Path]::GetTempPath()) -Filter RemoteIpMoProxy_International*_localhost_* | Remove-Item -Recurse -Force 
-			if ((Get-Command Import-Module).Parameters.ContainsKey('UseWindowsPowerShell'))
-			{
-				Import-Module -Name International -UseWindowsPowerShell -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -Force
-			}
-			else
-			{
-				Import-WinModule -Name International -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -Force
-			}
-
-			$inputLocale = @((New-WinUserLanguageList -Language $UserLocale).InputMethodTips)
-			$inputLocale += (New-WinUserLanguageList -Language 'en-us').InputMethodTips
-		}
+		$inputLocale = $languageList[$UserLocale]
+		$inputLocale += $languageList['en-us']
 	}
 	if ($inputLocale)
 	{